Player jumping and movement adjusted to be less floaty
The player takes turns more tightly and uses custom gravity forces to make the jump less floaty. Moving platform test script adjusted so it uses the SetPosition method of kinematic bodies. Bug found in which player is launched into the air when jumping just as the player collider clips the edge of another collider.